Inline `aircraft-shadow-layer` code
