+++ app/layers/index.js
@@ -1,2 +1,1 @@
 export { default as AircraftLayer } from './aircraft';
-export { default as AircraftShadowLayer } from './aircraft-shadow';
